// libs/enums/product.enum.ts
export enum ProductCategory {
	YACHT = 'YACHT',
	BOAT = 'BOAT',
	SNOWMOIBILE = 'SNOWMOBILE',
	ATV = 'ATV',
	JET_SKIS = 'JET_SKIS',
	TERRAIN_VEHICLES = 'TERRAIN_VEHICLES',
	ENGINES = 'ENGINES',
	OTHER = 'OTHER',
}

export enum ProductCondition {
	NEW = 'NEW',
	USED = 'USED',
}

export enum ProductStatus {
	ACTIVE = 'ACTIVE',
	SOLD = 'SOLD',
	DELETE = 'DELETE',
}

export enum ProductCurrency {
	USD = 'USD',
	KRW = 'KRW',
	EUR = 'EUR',
}

export enum InquiryItemType {
	PRODUCT = 'Product',
	EVENT = 'Event',
}

export enum ProductPriceType {
	FORSALE = 'FORSALE',
	RENT = 'RENT',
}

export enum ProductRentPeriod {
	HOUR = 'HOUR',
	DAY = 'DAY',
	WEEK = 'WEEK',
	MONTH = 'MONTH',
}